4 Player build. gonna start working on attacks now
