+++ Assets/Scripts/Player/PlayerOffense.cs
@@ -1,0 +1,26 @@
+﻿using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class PlayerOffense : MonoBehaviour {
+
+    //Modifiable Stats
+    private float AttackSpeed;
+    private float MultistrikeTimes;
+    private bool MultiStrike = false;
+    private float AoRRange = 0;
+    private float EnergyCostReduction = 0;
+    private float StatusEffectChance = 0;
+    private float LifeSteal = 0;
+    private float ProjectileSpeed = 0;
+    private float DamageIgnoreChance = 0;
+    // Use this for initialization
+    void Start () {
+		
+	}
+	
+	// Update is called once per frame
+	void Update () {
+		
+	}
+}
