+++ Assets/Scripts/Player/PlayerOffense.cs.meta
@@ -1,0 +1,13 @@
+fileFormatVersion: 2
+guid: 4027e5a51881708408f480469e33d41a
+timeCreated: 1535285943
+licenseType: Free
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
